Avoid signedness issues for pointer comparisons.

+++ source/fitz/load-tiff.c
@@ -882,7 +882,7 @@ tiff_read_tag(fz_context *ctx, struct tiff *tiff, unsigned offset)
 
 	case JPEGTables:
 		/* Check both value and value + count to allow for overflow */
-		if (value > tiff->ep - tiff->bp || value + count > tiff->ep - tiff->bp)
+		if (value > (size_t)(tiff->ep - tiff->bp) || value + count > (size_t)(tiff->ep - tiff->bp))
 			fz_throw(ctx, FZ_ERROR_GENERIC, "TIFF JPEG tables out of range");
 		tiff->jpegtables = tiff->bp + value;
 		tiff->jpegtableslen = count;
